qa_agent/requirements.txt
# Tier 1 Base Requirements
langgraph>=0.2.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
gradio>=4.0.0
pydantic>=2.0.0
pillow>=10.0.0
colorthief>=0.2.1
opencv-python-headless>=4.8.0
librosa>=0.10.0
mutagen>=1.47.0
faster-whisper>=0.10.0
semgrep>=1.45.0
ollama>=0.2.0
numpy>=1.24.0
scipy>=1.11.0
python-magic>=0.4.27

# Tier 2 Upgrade Requirements
tiktoken>=0.7.0
gitpython>=3.1.40
docker>=7.1.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
playwright>=1.44.0
pyannote-audio>=3.1.1
speechbrain>=1.0.0
keybert>=0.8.0
spacy>=3.7.0
praat-parselmouth>=0.4.3
